refactor(marquette): align application contract naming
