+++ crates/vize_marquette/src/validate.rs
@@ -1,14 +1,20 @@
 use std::collections::{BTreeMap, BTreeSet};
 
 use serde::{Deserialize, Serialize};
-use vize_carton::{String, ToCompactString};
+use vize_carton::String;
 
-use crate::{
-    ApplicationContract, CONTRACT_FORMAT_VERSION, EnvironmentConsumer, RenderingMode,
-    RuntimeFamily, Target,
+use crate::{ApplicationContract, CONTRACT_FORMAT_VERSION, EnvironmentConsumer, RuntimeFamily};
+
+mod rules;
+#[cfg(test)]
+mod tests;
+
+use rules::{
+    collect_unique_ids, contract_path, validate_capabilities, validate_environment_cycles,
+    validate_identifier, validate_rendering_target,
 };
 
-/// Severity of a platform-contract diagnostic.
+/// Severity of an application-marquette diagnostic.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Serialize, Deserialize)]
 #[serde(rename_all = "kebab-case")]
 pub enum DiagnosticSeverity {
@@ -61,16 +67,16 @@ pub fn validate_contract(contract: &ApplicationContract) -> Vec<ContractDiagnost
 
     if contract.format_version != CONTRACT_FORMAT_VERSION {
         diagnostics.push(ContractDiagnostic::error(
-            "VIZE_PLATFORM_001",
+            "VIZE_MARQUETTE_001",
             "formatVersion",
-            "unsupported platform contract format version",
+            "unsupported marquette format version",
         ));
     }
 
     validate_identifier(
         &contract.application,
         "application",
-        "VIZE_PLATFORM_002",
+        "VIZE_MARQUETTE_002",
         &mut diagnostics,
     );
 
@@ -83,19 +89,19 @@ pub fn validate_contract(contract: &ApplicationContract) -> Vec<ContractDiagnost
         validate_identifier(
             key,
             &contract_path("capabilities", key),
-            "VIZE_PLATFORM_003",
+            "VIZE_MARQUETTE_003",
             &mut diagnostics,
         );
         if key.as_str() != capability.id.as_str() {
             diagnostics.push(ContractDiagnostic::error(
-                "VIZE_PLATFORM_004",
+                "VIZE_MARQUETTE_004",
                 contract_path("capabilities", key),
                 "capability map key must equal capability.id",
             ));
         }
         if capability.version == 0 {
             diagnostics.push(ContractDiagnostic::error(
-                "VIZE_PLATFORM_005",
+                "VIZE_MARQUETTE_005",
                 contract_path("capabilities", key),
                 "capability version must be greater than zero",
             ));
@@ -127,7 +133,7 @@ pub fn validate_contract(contract: &ApplicationContract) -> Vec<ContractDiagnost
         let path = contract_path("environments", &environment.id);
         if !contract.targets.contains(&environment.target) {
             diagnostics.push(ContractDiagnostic::error(
-                "VIZE_PLATFORM_007",
+                "VIZE_MARQUETTE_007",
                 path.clone(),
                 "environment target must be declared in targets",
             ));
@@ -135,13 +141,13 @@ pub fn validate_contract(contract: &ApplicationContract) -> Vec<ContractDiagnost
         for dependency in &environment.depends_on {
             if dependency.as_str() == environment.id.as_str() {
                 diagnostics.push(ContractDiagnostic::error(
-                    "VIZE_PLATFORM_008",
+                    "VIZE_MARQUETTE_008",
                     path.clone(),
                     "environment cannot depend on itself",
                 ));
             } else if !environment_ids.contains(dependency) {
                 diagnostics.push(ContractDiagnostic::error(
-                    "VIZE_PLATFORM_009",
+                    "VIZE_MARQUETTE_009",
                     path.clone(),
                     "environment dependency does not exist",
                 ));
@@ -161,7 +167,7 @@ pub fn validate_contract(contract: &ApplicationContract) -> Vec<ContractDiagnost
             )
         {
             diagnostics.push(ContractDiagnostic::warning(
-                "VIZE_PLATFORM_010",
+                "VIZE_MARQUETTE_010",
                 path,
                 "client environment uses a server-oriented runtime; declare an adapter capability if this is intentional",
             ));
@@ -175,7 +181,7 @@ pub fn validate_contract(contract: &ApplicationContract) -> Vec<ContractDiagnost
         if let Some(environment) = &backend.environment {
             if !environment_ids.contains(environment) {
                 diagnostics.push(ContractDiagnostic::error(
-                    "VIZE_PLATFORM_011",
+                    "VIZE_MARQUETTE_011",
                     path.clone(),
                     "backend environment does not exist",
                 ));
@@ -186,7 +192,7 @@ pub fn validate_contract(contract: &ApplicationContract) -> Vec<ContractDiagnost
                 .is_some_and(|candidate| candidate.consumer != EnvironmentConsumer::Server)
             {
                 diagnostics.push(ContractDiagnostic::error(
-                    "VIZE_PLATFORM_012",
+                    "VIZE_MARQUETTE_012",
                     path.clone(),
                     "backend environment must be a server consumer",
                 ));
@@ -204,7 +210,7 @@ pub fn validate_contract(contract: &ApplicationContract) -> Vec<ContractDiagnost
         let path = contract_path("protocols", &protocol.id);
         if !backend_ids.contains(&protocol.backend) {
             diagnostics.push(ContractDiagnostic::error(
-                "VIZE_PLATFORM_013",
+                "VIZE_MARQUETTE_013",
                 path.clone(),
                 "protocol backend does not exist",
             ));
@@ -222,14 +228,14 @@ pub fn validate_contract(contract: &ApplicationContract) -> Vec<ContractDiagnost
         let path = contract_path("routes", &route.id);
         if !route.path.starts_with('/') {
             diagnostics.push(ContractDiagnostic::error(
-                "VIZE_PLATFORM_014",
+                "VIZE_MARQUETTE_014",
                 path.clone(),
                 "route path must start with /",
             ));
         }
         if !environment_ids.contains(&route.environment) {
             diagnostics.push(ContractDiagnostic::error(
-                "VIZE_PLATFORM_015",
+                "VIZE_MARQUETTE_015",
                 path.clone(),
                 "route environment does not exist",
             ));
@@ -238,7 +244,7 @@ pub fn validate_contract(contract: &ApplicationContract) -> Vec<ContractDiagnost
             && !backend_ids.contains(backend)
         {
             diagnostics.push(ContractDiagnostic::error(
-                "VIZE_PLATFORM_016",
+                "VIZE_MARQUETTE_016",
                 path.clone(),
                 "route backend does not exist",
             ));
@@ -246,7 +252,7 @@ pub fn validate_contract(contract: &ApplicationContract) -> Vec<ContractDiagnost
         if let Some(protocol) = &route.protocol {
             if !protocol_ids.contains(protocol) {
                 diagnostics.push(ContractDiagnostic::error(
-                    "VIZE_PLATFORM_017",
+                    "VIZE_MARQUETTE_017",
                     path.clone(),
                     "route protocol does not exist",
                 ));
@@ -258,7 +264,7 @@ pub fn validate_contract(contract: &ApplicationContract) -> Vec<ContractDiagnost
                     .is_some_and(|candidate| candidate.backend.as_str() != route_backend.as_str())
             {
                 diagnostics.push(ContractDiagnostic::error(
-                    "VIZE_PLATFORM_018",
+                    "VIZE_MARQUETTE_018",
                     path.clone(),
                     "route protocol and backend must refer to the same service",
                 ));
@@ -275,7 +281,7 @@ pub fn validate_contract(contract: &ApplicationContract) -> Vec<ContractDiagnost
         let key = (&route.environment, &route.path);
         if let Some(previous) = route_paths.insert(key, &route.id) {
             diagnostics.push(ContractDiagnostic::error(
-                "VIZE_PLATFORM_019",
+                "VIZE_MARQUETTE_019",
                 path,
                 previous.clone(),
             ));
@@ -289,7 +295,7 @@ pub fn validate_contract(contract: &ApplicationContract) -> Vec<ContractDiagnost
             .any(|environment| &environment.target == target)
         {
             diagnostics.push(ContractDiagnostic::warning(
-                "VIZE_PLATFORM_020",
+                "VIZE_MARQUETTE_020",
                 "targets",
                 "declared target has no environment",
             ));
@@ -300,231 +306,4 @@ pub fn validate_contract(contract: &ApplicationContract) -> Vec<ContractDiagnost
         (&left.path, left.code, &left.message).cmp(&(&right.path, right.code, &right.message))
     });
     diagnostics
-}
-
-fn collect_unique_ids<'a>(
-    collection: &str,
-    ids: impl Iterator<Item = &'a String>,
-    diagnostics: &mut Vec<ContractDiagnostic>,
-) -> BTreeSet<String> {
-    let mut unique = BTreeSet::new();
-    for id in ids {
-        validate_identifier(
-            id,
-            &contract_path(collection, id),
-            "VIZE_PLATFORM_006",
-            diagnostics,
-        );
-        if !unique.insert(id.clone()) {
-            diagnostics.push(ContractDiagnostic::error(
-                "VIZE_PLATFORM_006",
-                contract_path(collection, id),
-                "identifier must be unique within its collection",
-            ));
-        }
-    }
-    unique
-}
-
-fn validate_identifier(
-    id: &str,
-    path: &str,
-    code: &'static str,
-    diagnostics: &mut Vec<ContractDiagnostic>,
-) {
-    let mut characters = id.bytes();
-    let valid_first = characters
-        .next()
-        .is_some_and(|value| value.is_ascii_lowercase() || value.is_ascii_digit());
-    let valid_rest = characters.all(|value| {
-        value.is_ascii_lowercase() || value.is_ascii_digit() || matches!(value, b'-' | b'_' | b'.')
-    });
-    if !valid_first || !valid_rest {
-        diagnostics.push(ContractDiagnostic::error(
-            code,
-            path,
-            "identifier must use lowercase ASCII letters, digits, dash, underscore, or dot",
-        ));
-    }
-}
-
-fn validate_capabilities(
-    path: &str,
-    required: &BTreeSet<String>,
-    declared: &BTreeSet<String>,
-    diagnostics: &mut Vec<ContractDiagnostic>,
-) {
-    for capability in required {
-        if !declared.contains(capability) {
-            diagnostics.push(ContractDiagnostic::error(
-                "VIZE_PLATFORM_021",
-                path,
-                "referenced capability is not declared",
-            ));
-        }
-    }
-}
-
-fn validate_environment_cycles(
-    contract: &ApplicationContract,
-    diagnostics: &mut Vec<ContractDiagnostic>,
-) {
-    let graph = contract
-        .environments
-        .iter()
-        .map(|environment| (&environment.id, &environment.depends_on))
-        .collect::<BTreeMap<_, _>>();
-    let mut visiting = BTreeSet::new();
-    let mut visited = BTreeSet::new();
-
-    for id in graph.keys() {
-        if has_cycle(id, &graph, &mut visiting, &mut visited) {
-            diagnostics.push(ContractDiagnostic::error(
-                "VIZE_PLATFORM_022",
-                contract_path("environments", id),
-                "environment dependency graph must be acyclic",
-            ));
-        }
-    }
-}
-
-fn has_cycle<'a>(
-    id: &'a String,
-    graph: &BTreeMap<&'a String, &'a BTreeSet<String>>,
-    visiting: &mut BTreeSet<String>,
-    visited: &mut BTreeSet<String>,
-) -> bool {
-    if visited.contains(id) {
-        return false;
-    }
-    if !visiting.insert(id.clone()) {
-        return true;
-    }
-    let cyclic = graph.get(id).is_some_and(|dependencies| {
-        dependencies.iter().any(|dependency| {
-            graph
-                .get_key_value(dependency)
-                .is_some_and(|(canonical, _)| has_cycle(canonical, graph, visiting, visited))
-        })
-    });
-    visiting.remove(id);
-    visited.insert(id.clone());
-    cyclic
-}
-
-fn validate_rendering_target(
-    contract: &ApplicationContract,
-    route: &crate::Route,
-    path: &str,
-    diagnostics: &mut Vec<ContractDiagnostic>,
-) {
-    let Some(environment) = contract
-        .environments
-        .iter()
-        .find(|environment| environment.id == route.environment)
-    else {
-        return;
-    };
-    let valid = match route.rendering {
-        RenderingMode::Native => environment.target == Target::Native,
-        RenderingMode::Desktop => environment.target == Target::Desktop,
-        RenderingMode::Terminal => environment.target == Target::Terminal,
-        RenderingMode::Client
-        | RenderingMode::Static
-        | RenderingMode::Server
-        | RenderingMode::Stream
-        | RenderingMode::Partial
-        | RenderingMode::Hybrid => environment.target == Target::Web,
-    };
-    if !valid {
-        diagnostics.push(ContractDiagnostic::error(
-            "VIZE_PLATFORM_023",
-            path,
-            "rendering mode is not compatible with the route environment target",
-        ));
-    }
-}
-
-fn contract_path(collection: &str, id: &str) -> String {
-    let mut path = collection.to_compact_string();
-    path.push('.');
-    path.push_str(id);
-    path
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::{
-        Backend, BackendFamily, Environment, Protocol, ProtocolFamily, RenderingMode, Route,
-        RuntimeFamily,
-    };
-
-    #[test]
-    fn accepts_a_cross_language_web_contract() {
-        let mut contract = ApplicationContract::new("shop");
-        contract.targets.insert(Target::Web);
-        contract.environments.push(Environment::new(
-            "client",
-            Target::Web,
-            EnvironmentConsumer::Client,
-            RuntimeFamily::Browser,
-        ));
-        contract.environments.push(Environment::new(
-            "server",
-            Target::Web,
-            EnvironmentConsumer::Server,
-            RuntimeFamily::Rust,
-        ));
-        contract
-            .backends
-            .push(Backend::new("api", BackendFamily::Rust).with_environment("server"));
-        contract.protocols.push(Protocol::new(
-            "api.query",
-            ProtocolFamily::SchemaQuery,
-            "api",
-        ));
-        let mut route = Route::new("home", "/", "client", RenderingMode::Hybrid);
-        route.backend = Some("api".into());
-        route.protocol = Some("api.query".into());
-        contract.routes.push(route);
-
-        assert_eq!(validate_contract(&contract), Vec::new());
-    }
-
-    #[test]
-    fn reports_cycles_and_cross_reference_errors() {
-        let mut contract = ApplicationContract::new("Broken App");
-        contract.targets.insert(Target::Web);
-        let mut first = Environment::new(
-            "first",
-            Target::Web,
-            EnvironmentConsumer::Server,
-            RuntimeFamily::JavaScript,
-        );
-        first.depends_on.insert("second".into());
-        let mut second = Environment::new(
-            "second",
-            Target::Web,
-            EnvironmentConsumer::Server,
-            RuntimeFamily::JavaScript,
-        );
-        second.depends_on.insert("first".into());
-        contract.environments.extend([first, second]);
-        contract.routes.push(Route::new(
-            "home",
-            "missing-slash",
-            "missing",
-            RenderingMode::Native,
-        ));
-
-        let codes = validate_contract(&contract)
-            .into_iter()
-            .map(|diagnostic| diagnostic.code)
-            .collect::<BTreeSet<_>>();
-        assert!(codes.contains("VIZE_PLATFORM_002"));
-        assert!(codes.contains("VIZE_PLATFORM_014"));
-        assert!(codes.contains("VIZE_PLATFORM_015"));
-        assert!(codes.contains("VIZE_PLATFORM_022"));
-    }
 }
