+++ tests/fixtures/marquette/invalid.expected.json
@@ -1,0 +1,7 @@
+[
+  "VIZE_MARQUETTE_002",
+  "VIZE_MARQUETTE_009",
+  "VIZE_MARQUETTE_013",
+  "VIZE_MARQUETTE_015",
+  "VIZE_MARQUETTE_017"
+]
