--- tests/fixtures/platform-contract/invalid.expected.json
@@ -1,7 +1,0 @@
-[
-  "VIZE_PLATFORM_002",
-  "VIZE_PLATFORM_009",
-  "VIZE_PLATFORM_013",
-  "VIZE_PLATFORM_015",
-  "VIZE_PLATFORM_017"
-]
